pyproject.toml
[tool.poetry]
name = "testpythonbot"
version = "0.1.0"
description = ""
authors = ["wnkbll <arzhevitin.g@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.10.0"
redis = {extras = ["hiredis"], version = "^5.0.7"}
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
apscheduler = "^3.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
